+++ src/common/replayBrowser/index.ts
@@ -1,3 +1,3 @@
 export * from "./types";
 export * from "./folderTree";
-export * from "./fileLoader";
+export * from "./loadFolder";
